Move to /track endpoint to avoid conflicts
